Upload of my current work V3
V3 work for LSTM multilayers. I choose to keep the tensorflow version, stop with the keras version (seems to be too magic box and less intuitive for global understanding). The model seems to work for any number of hidden layer, timestep or neurons in layers.

But for now layers need to have same size (same number of neurons)